refactoring position
