+++ Sprint1/unibo.mocktruck/resources/observers/coapobs/CoapTestObserver.kt
@@ -1,4 +1,4 @@
-package test.it.unibo.coapobs
+package observers.coapobs
 
 import org.eclipse.californium.core.CoapHandler
 import org.eclipse.californium.core.CoapResponse
